mv generateMeme into its own meme svc
